nice version ;)

+++ ft_fdf/srcs/draw/color2.c
@@ -6,9 +6,10 @@
 /*   By: vmeessen <vmeessen@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/14 12:40:08 by vmeessen          #+#    #+#             */
-/*   Updated: 2025/04/14 13:45:00 by vmeessen         ###   ########.fr       */
+/*   Updated: 2025/04/15 15:07:22 by vmeessen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+
 
 int	interpolate_color(int colr1, int colr2, double ratio)
 {
@@ -22,16 +23,17 @@ int	interpolate_color(int colr1, int colr2, double ratio)
 	return ((red << 16) | (green << 8) | blue);
 }
 
+
 /*
 Donne l interpolation entre deux couleur par rapport a la hauteur
 */
-int	calculate_color(int z, int min_z, int max_z)
+int	calculate_color(long z, int min_z, int max_z)
 {
 	double	ratio;
 
 	if (max_z == min_z)
 		ratio = 0;
 	else
-		ratio = (double)(z - min_z) / (max_z - min_z);
+		ratio = (double)(z - (long)min_z) / (double)((long)max_z - (long)min_z);
 	return (interpolate_color(0x00FF00, 0xFF00FF, ratio));
 }
